pyproject.toml
[tool.poetry]
name = "carma-usage-scraper"
version = "0.1.0"
description = "Scrape electricity usage data from Carma Smart Metering"
authors = ["Charlie Wang <charliewang@fastmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
playwright = "^1.12.1"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
flake8 = "^3.9.2"
mypy = "^0.910"
isort = "^5.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.mypy_cache
    | \.venv
    | _build
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
